+++ my-frontend/src/constants/Routes.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import {Route, Routes  } from 'react-router-dom'; // Import Routes and Route from react-router-dom
-import Home from './components/intakes/Home';
-import Intakes from './components/intakes/Intakes';
-import Courses from './components/courses/Courses';
-import AddCourse from './components/courses/AddCourse';
-import AddIntake from './components/intakes/AddIntake';
-import EditIntake from './components/intakes/EditIntake';
-import Sample from './components/intakes/Sample';
+import Home from '../components/intakes/Home';
+import Intakes from '../components/intakes/Intakes';
+import Courses from '../components/courses/Courses';
+import AddCourse from '../components/courses/AddCourse';
+import AddIntake from '../components/intakes/AddIntake';
+import EditIntake from '../components/intakes/EditIntake';
+import Sample from '../components/intakes/Sample';
 
 const AppRoutes: React.FC = () => {
   return (
