+++ my-frontend/src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from 'react-router-dom';
-import { ThemeProviderWrapper } from './Theme2';
+import { ThemeProviderWrapper } from './constants/ThemeConstants';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
